+++ System/Component.hpp
@@ -19,7 +19,7 @@ namespace System {
 
     class Component {
         protected:
-            short _uniqueID = 0;
+            int _uniqueID = 0;
             short _orderOfExecution =  0;
 
         public:
@@ -31,7 +31,7 @@ namespace System {
 
             virtual void setOrderOfExecution(const short& newOrderOfExecution) = 0;
 
-            virtual short getUniqueID() = 0;
+            virtual int getUniqueID() = 0;
 
             virtual std::string toString() = 0;
